% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fd_indices.R
\name{calculate_fd_indices}
\alias{calculate_fd_indices}
\title{Calculate Functional Diversity Indices}
\usage{
calculate_fd_indices(trait_matrix, abundance_matrix)
}
\arguments{
\item{trait_matrix}{DataFrame containing species traits}

\item{abundance_matrix}{DataFrame containing species abundance by site}
}
\value{
List of FD indices (FRic, FEve, FDiv)
}
\description{
Calculate Functional Diversity Indices
}
